Refactor and delete unused code
