+++ Sources/ApodiniLoggingSupport/Logger.MetadataValue+Codable.swift
@@ -1,0 +1,47 @@
+//
+// This source file is part of the Apodini open source project
+//
+// SPDX-FileCopyrightText: 2019-2021 Paul Schmiedmayer and the Apodini project authors (see CONTRIBUTORS.md) <paul.schmiedmayer@tum.de>
+//
+// SPDX-License-Identifier: MIT
+//
+
+import Foundation
+import Logging
+
+/// Make ``Logger.MetadataValue`` ``Codable``
+extension Logger.MetadataValue: Codable {
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        
+        switch self {
+        case let .string(string):
+            try container.encode(string)
+        case let .stringConvertible(stringConvertible):
+            try container.encode(stringConvertible.description)
+        case let .dictionary(dictionary):
+            try container.encode(dictionary)
+        case let .array(array):
+            try container.encode(array)
+        }
+    }
+    
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.singleValueContainer()
+        
+        if container.decodeNil() {
+            self = .string("null")
+        } else if let string = try? container.decode(String.self) {
+            self = .string(string)
+        } else if let array = try? container.decode([Logger.MetadataValue].self) {
+            self = .array(array)
+        } else if let dictionary = try? container.decode(Logger.Metadata.self) {
+            self = .dictionary(dictionary)
+        } else {
+            throw DecodingError.dataCorrupted(
+                DecodingError.Context(codingPath: decoder.codingPath,
+                                      debugDescription: "Encountered unexpected JSON values")
+            )
+        }
+    }
+}
